==> 2.4 species area.rmd <==
---
title: "2.4 species area"
author: "TG"
date: "13 9 2021"
output:
  html_document:
    df_print: paged
  prettydoc::html_pretty:
    theme: cayman
    highlight: github
---
 

```{r}
load("fish_full_data.rdata")
select <- dplyr::select
```

```{r message=FALSE, warning=FALSE}
library(dplyr)
library(tidyr)
library(lubridate)
library(ggplot2)
library(stringr)
library(Hmisc)
library(vegan)
library(mobr)
library(lme4)
library(lmerTest)
library(MuMIn)

fish_k<-c("#669bbc","#336683","#1a4b66","#003049","#ed3140","#c1121f","#710912")
fish_all<-c("#669bbc","#336683","#003049","#ed3140","#c1121f","#710912")

```


### general steps:

1) filter confidence = 1

for the years 2015-2017 where there is no confidance value i set the value 0

```{r}

sp_area<-fish_full_data
sp_area$Confidence[is.na(sp_area$Confidence)]<-(0)
sp_area<- sp_area %>% filter(Confidence < 1)

```


2) remove irregular sites


```{r}

knolls_to_remove<-c("South bridge reserve","North bridge reserve","south bridge reserve",
                    "Tamar","north bridge reserve")

sp_area<-sp_area %>% filter(!Knoll %in% knolls_to_remove) %>% filter(Site !="North beach")

```

3) keep one observer

```{r}
one_observer<-list()
loop<-1
set.seed(1)

for (i in unique(sp_area$survey_id)) {
  
  data <- sp_area %>% filter(survey_id == i)
  
  obs_amount <- length(unique(data$Observer))
  obs_name <- unique(data$Observer)
  chosen_obs <- sample(obs_name,1)  
 
  filter_data<-data %>% filter(Observer == chosen_obs)
  one_observer[[loop]]<-filter_data
  loop<-loop+1
  
  
}

one_observer<- bind_rows(one_observer)

rm(data,filter_data)
```



```{r}
sp_list<-fish_full_data %>% distinct(Species,.keep_all = T) %>% select(Species,Species_2015,Family,Status) %>% arrange(Family,Species)

```


the connection between resident fish and surface area


```{r}
only_cryptic <- one_observer %>% filter(Status != "T")

```


3) set the number of species per knoll


```{r}

species_knoll<- only_cryptic %>% group_by(survey_id) %>% mutate(Species_number = n_distinct(Species_2015)) %>% select(-Species,-Species_2015,-Family,-Abundance,-Length,-Confidence,-Status,-a,-b) %>%  ungroup() %>% distinct(.keep_all = T)

```

* ill leave 2007 where only kazza was sampled out

```{r}
species_knoll<-species_knoll %>% filter(Year != 2017)


```

add month 

```{r}
species_knoll$month<-month(species_knoll$Date)
species_knoll$month<-as.factor(species_knoll$month)

species_knoll$storm<-as.factor(species_knoll$storm)
species_knoll$Season <- ifelse(species_knoll$month == 6,"Spring","Fall")

species_knoll$Year<-as.factor(species_knoll$Year)
species_knoll$Site<-as.factor(species_knoll$Site)
```

models

```{r}
str(species_knoll)


model_list<-list()


```

## model 1 - set - decide how to model the year/season/storm


```{r}
model_list[[1]]<-lmer(formula = log(Species_number) ~ log(Surface_Area)*Site +(1|year_month) ,data = species_knoll)

summary(model_list[[1]])
r.squaredGLMM(model_list[[1]])

model_list[[1]]<-lm(formula = log(Species_number) ~ log(Surface_Area)*Site + year_month ,data = species_knoll)


model_list[[2]]<-lmer(formula = log(Species_number) ~ log(Surface_Area)*Site +(1|Year)+Season ,data = species_knoll)


summary(model_list[[2]])
r.squaredGLMM(model_list[[2]])

model_list[[3]]<-lmer(formula = log(Species_number) ~ log(Surface_Area)*Site +(1|Year) +Season + storm ,data = species_knoll)


summary(model_list[[3]])
r.squaredGLMM(model_list[[3]])


model_list[[4]]<-lmer(formula = log(Species_number) ~ log(Surface_Area)*Site +(1|Year)  + storm ,data = species_knoll)


summary(model_list[[4]])
r.squaredGLMM(model_list[[4]])

model_list[[5]]<-lmer(formula = log(Species_number) ~ log(Surface_Area)*Site +(1|year_month)  + storm ,data = species_knoll)


summary(model_list[[5]])
r.squaredGLMM(model_list[[5]])

aic_table<-model.sel(model_list)

```


lets keep it simple and keep going with model 1 as the base model and test the coral related variable



```{r}
species_knoll<-na.omit(species_knoll)
```


```{r}
model_2<-list()


model_2[[1]]<-lmer(formula = log(Species_number) ~ log(Surface_Area)*Site +(1|year_month) ,data = species_knoll)



model_2[[2]]<-lmer(formula = log(Species_number) ~ log(Surface_Area)*Site +(1|year_month)+Max_depth ,data = species_knoll)



model_2[[3]]<-lmer(formula = log(Species_number) ~ log(Surface_Area)*Site +(1|year_month)+Max_depth +Hight ,data = species_knoll)




model_2[[4]]<-lmer(formula = log(Species_number) ~ log(Surface_Area)*Site +(1|year_month)+Max_depth + Hight +coral_cover ,data = species_knoll)




model_2[[5]]<-lmer(formula = log(Species_number) ~ log(Surface_Area)*Site +(1|year_month)+Max_depth + Hight +coral_cover +Algae ,data = species_knoll)




model_2[[6]]<-lmer(formula = log(Species_number) ~ log(Surface_Area)*Site +(1|year_month)+Max_depth + Hight  +Algae ,data = species_knoll)




model_2[[7]]<-lmer(formula = log(Species_number) ~ log(Surface_Area)*Site +(1|year_month)+Max_depth + Algae ,data = species_knoll)




model_2[[8]]<-lmer(formula = log(Species_number) ~ log(Surface_Area)*Site +(1|year_month)+Min_depth ,data = species_knoll)



aic_st2<-model.sel(model_2)

summary(model_2[[2]])
r.squaredGLMM(model_2[[2]])

```


in the second step, model 2 is the best!

lmer(formula = log(Species_number) ~ log(Surface_Area)*Site +(1|year_month)+Max_depth ,data = species_knoll)
```{r}
sp_ar_model<-model_2[[2]]

plot(sp_ar_model)
```


```{r}
surface_area = rep(seq(from = 1,to = 45,by = 1),24) # Length = 45
site = rep(unique(species_knoll$Site),each = 270) # Length = 4
year = rep(rep(unique(species_knoll$year_month),each = 45),4) # Length = 6




mean(species_knoll$Max_depth)
median(species_knoll$Max_depth)
  
predict_cryptic<-data.frame(Site = site,
                       year_month = year,
                       Surface_Area= surface_area,
                       Max_depth = rep(4.6))




predict_cryptic$Surface_Area_log<-log(predict_cryptic$Surface_Area)

predict_species_cryptic<-predict(sp_ar_model,predict_cryptic,type="response")

cyptic_predicted<-cbind(predict_cryptic,predict_species_cryptic)

cyptic_predicted$site_year<-paste(cyptic_predicted$Site,cyptic_predicted$year_month,sep="_")

cyptic_predicted$species_number<-exp(cyptic_predicted$predict_species_cryptic)


cyptic_predicted$year_month<- as.factor(cyptic_predicted$year_month)


ggplot(data = cyptic_predicted,aes(x= Surface_Area,y = species_number,color= site))+
  theme_classic()+geom_smooth()+geom_point()




ggplot(data = cyptic_predicted,aes(x= Surface_Area,y = species_number,color= year_month))+
  theme_classic()+geom_smooth()+geom_point(data = species_knoll,aes(Surface_Area,Species_number,color = year_month),alpha = 0.3)+
  ylab("Species number")+
  xlab(bquote('Knoll surface area '~(m^2)))+
  theme(axis.text.x = element_text(size=16),
        axis.text.y = element_text(size =16),
        axis.title.x = element_text(size=18),
        axis.title.y = element_text(size=18),
        legend.title = element_text(size=18),
        legend.text = element_text(size=16),
        strip.text.x = element_text(size = 16))+
  scale_color_manual(values=fish_all)




ggplot(data = cyptic_predicted,aes(x= Surface_Area,y = species_number,color= site))+
geom_line()+ facet_grid(year_month ~ .)
```


# filter the june dates and check the plots


```{r}
no_spring<-species_knoll

no_spring<-no_spring %>% filter(Season != "Spring")

no_spring$Year<-droplevels(no_spring$Year)

model_no_spring<-lmer(formula = log(Species_number) ~ log(Surface_Area)*Site +(1|Year)+Max_depth ,data = no_spring)

summary(model_no_spring)
r.squaredGLMM(model_no_spring)


surface_area = rep(seq(from = 1,to = 45,by = 1),16) # Length = 45
site = rep(unique(species_knoll$Site),each = 180) # Length = 4
year = rep(rep(unique(no_spring$Year),each = 45),4) # Length = 4


surface_area = rep(seq(from = 1,to = 60,by = 1),16) # Length = 45
site = rep(unique(species_knoll$Site),each = 240) # Length = 4
year = rep(rep(unique(no_spring$Year),each = 60),4) # Length = 4

predict_no_spring<-data.frame(Site = site,
                       Year = year,
                       Surface_Area= surface_area,
                       Max_depth = rep(4.6))




predict_no_spring$Surface_Area_log<-log(predict_no_spring$Surface_Area)

predict_species_cryptic_ns<-predict(model_no_spring,predict_no_spring,type="response")

predict_no_spring<-cbind(predict_no_spring,predict_species_cryptic_ns)

predict_no_spring$site_year<-paste(predict_no_spring$Site,predict_no_spring$Year,sep="_")

predict_no_spring$species_number<-exp(predict_no_spring$predict_species_cryptic_ns)


predict_no_spring$Year<- as.factor(predict_no_spring$Year)


ggplot(data = predict_no_spring,aes(x= Surface_Area,y = species_number,color= site))+
  theme_classic()+geom_smooth()+geom_point()


ggplot(data = predict_no_spring,aes(x= Surface_Area,y = species_number,color= site))+
  theme_classic()+geom_smooth()


ggplot(data = predict_no_spring,aes(x= Surface_Area,y = species_number,color= site))+
geom_line()+ facet_grid(Year ~ .)



```

# add the real data point to the plot...


```{r}
ggplot(data = predict_no_spring,aes(x= Surface_Area,y = species_number,color= site))+
  theme_classic()+geom_smooth()+
  geom_point(data = no_spring,aes(Surface_Area,Species_number,color = Site),alpha = 0.3)


# add the after storm point (outside the model...)


only_storm<- species_knoll %>% filter(year_month=="2020 a")

ggplot(data = predict_no_spring,aes(x= Surface_Area,y = species_number,color= site))+
  theme_classic()+geom_smooth()+
  geom_point(data = no_spring,aes(Surface_Area,Species_number,color = Site),alpha = 0.3)+
  geom_point(data = only_storm,aes(Surface_Area,Species_number),
             color = "black",alpha = 0.7,shape=17)+
  ylab("Species number")+
  xlab(bquote('Knoll surface area '~(m^2)))+
  theme(axis.text.x = element_text(size=16),
        axis.text.y = element_text(size =16),
        axis.title.x = element_text(size=18),
        axis.title.y = element_text(size=18),
        legend.title = element_text(size=18),
        legend.text = element_text(size=16),
        strip.text.x = element_text(size = 16))

plot_path<-"C:\\Users\\TalG\\Documents\\PhD\\Eco camp\\R Scripts\\plots nov 2021"

#ggsave("species_area.pdf", width = 10, height = 8, dpi = 300,path = plot_path)
#ggsave("species_area.png", width = 10, height = 8, dpi = 300,path = plot_path)

```


```{r}
no_nr_outliar<-species_knoll %>% filter(knoll_id != "NR_11",knoll_id != "NR_17")

model_no_outliar<-lmer(formula = log(Species_number) ~ log(Surface_Area)*Site +(1|year_month)+Max_depth ,data = no_nr_outliar)



surface_area = rep(seq(from = 1,to = 60,by = 1),16) # Length = 45
site = rep(unique(species_knoll$Site),each = 240) # Length = 4
year = rep(rep(unique(no_spring$year_month),each = 60),4) # Length = 4

predict_no_out<-data.frame(Site = site,
                       year_month = year,
                       Surface_Area= surface_area,
                       Max_depth = rep(4.6))




predict_no_out$Surface_Area_log<-log(predict_no_spring$Surface_Area)


predict_species_no_out<-predict(model_no_outliar,predict_no_out,type="response")

predict_no_out<-cbind(predict_no_out,predict_species_no_out)

predict_no_out$site_year<-paste(predict_no_out$Site,predict_no_out$year_month,sep="_")

predict_no_out$species_number<-exp(predict_no_out$predict_species_no_out)


predict_no_out$year_month<- as.factor(predict_no_out$year_month)

ggplot(data = predict_no_out,aes(x= Surface_Area,y = species_number,color= site))+
  theme_classic()+geom_smooth()+
  geom_point(data = no_nr_outliar,aes(Surface_Area,Species_number,color = Site),alpha = 0.3)+
  ylab("Species number")+
  xlab(bquote('Knoll surface area '~(m^2)))+
  theme(axis.text.x = element_text(size=16),
        axis.text.y = element_text(size =16),
        axis.title.x = element_text(size=18),
        axis.title.y = element_text(size=18),
        legend.title = element_text(size=18),
        legend.text = element_text(size=16),
        strip.text.x = element_text(size = 16))



```
