==> 2.7 fish shannon area.rmd <==
---
title: "2.7 fish shannon area"
author: "TG"
date: "1 2 2022"
output: html_document
---

```{r}
load("fish_full_data.rdata")
load("survey_list.Rdata")
```

```{r message=FALSE, warning=FALSE}
library(dplyr)
library(tidyr)
library(lubridate)
library(ggplot2)
library(stringr)
library(Hmisc)
library(vegan)
library(mobr)
#library(sars)
library(lme4)
library(lmerTest)
library(MuMIn)



fish_k<-c("#669bbc","#336683","#1a4b66","#003049","#ed3140","#c1121f","#710912")
fish_all<-c("#669bbc","#336683","#003049","#ed3140","#c1121f","#710912")
color<-c("#56dde0", "#33658a", "#f5af2d","#f2641a")
```

1) set one observer

```{r}
one_observer<-list()
loop<-1
set.seed(1)
for (i in unique(fish_full_data$survey_id)) {
  
  data <- fish_full_data %>% filter(survey_id == i)
  
  obs_amount <- length(unique(data$Observer))
  obs_name <- unique(data$Observer)
  chosen_obs <- sample(obs_name,1)  
 
  filter_data<-data %>% filter(Observer == chosen_obs)
  one_observer[[loop]]<-filter_data
  loop<-loop+1
  
  
}

one_observer<- bind_rows(one_observer)

rm(data,filter_data)
```


2) filter confidence = 0

for the years 2015-2017 where there is no confidance value i set the value 0

```{r}

one_observer$Confidence[is.na(one_observer$Confidence)]<-(0)
one_observer<-one_observer %>% filter(Confidence <1)

```

3) filter the north beach
```{r}

one_observer<-one_observer %>%
  filter(Site !="North beach",
         ! Knoll %in% c("Tamar","South bridge reserve","south bridge reserve",
                        "North bridge reserve","north bridge reserve"))
  
```



4) add the month to the sample

```{r include=FALSE}

one_observer$year_month<-ifelse(month(one_observer$Date)>6,
                                paste(one_observer$Year,"b",sep= " "),
                                paste(one_observer$Year,"a",sep= " "))



only_cryptic <- one_observer %>% filter(Status != "T")
```





create wide data for:

1. species by 2015 categories - wide_2015 (for each site comparison)

2. species by each year resolution (i.e Species column) - wide_data (for each year comparison)




### 1. mobr_2015_data

species by 2015 categories - wide_2015 (for each site comparison)


```{r include=FALSE}
all_fish_2015 <- only_cryptic %>%
  group_by(Site,Year,year_month,Knoll,DD_X,DD_Y,Surface_Area,Max_depth,Date,survey_id,Species_2015) %>%
  summarise("aboundance"=sum(Abundance))


wide_2015<-spread(all_fish_2015,Species_2015,aboundance,fill = 0)


```

### 2. wide data for 2018-2021

species by each year resolution (i.e Species column) - wide_data (for each year comparison)


```{r include=FALSE}
all_fish <- only_cryptic %>%
  group_by(Site,Year,year_month,Knoll,DD_X,DD_Y,Surface_Area,Max_depth,Date,survey_id,Species) %>%
  summarise("aboundance"=sum(Abundance))


wide_data<-spread(all_fish,Species,aboundance,fill = 0)

wide_data<-wide_data %>% filter(Year >2017)

```

add the shannon values to the data 2015


```{r}
wide_2015$shannon_div<-diversity(wide_2015[,11:ncol(wide_2015)],index="shannon")

wide_2015<-wide_2015 %>%
  relocate(shannon_div,.after=survey_id) %>% 
  mutate("shannon_effective"=exp(shannon_div),.after=shannon_div)
  
wide_2015<- wide_2015[,1:12]

wide_2015<-wide_2015 %>% filter(Year !=2017)

wide_2015$year_month<-as.factor(wide_2015$year_month)

```


add the Shannon values to the data 2018-2021


```{r}
wide_data$shannon_div<-diversity(wide_data[,11:ncol(wide_data)],index="shannon")

wide_data<-wide_data %>%
  relocate(shannon_div,.after=survey_id) %>% 
  mutate("shannon_effective"=exp(shannon_div),.after=shannon_div)
  
wide_data<- wide_data[,1:12]

wide_data$year_month<-as.factor(wide_data$year_month)
```


# 2015 data models


```{r}

model_2015<-lm(formula = log(shannon_effective) ~ log(Surface_Area)*year_month+Site+Max_depth ,data = wide_2015)
summary(model_2015)



model_5_gg=ggeffects::ggpredict(model_2015,terms = c('Surface_Area[1:66 by=1]','year_month'))

ggplot()+
  theme_classic()+
  geom_line(data = model_5_gg,aes(x= log(x),y = log(predicted),color= group),size=1.5)+ 
  geom_point(data = wide_2015,aes(log(Surface_Area),log(shannon_effective),color = year_month),alpha = 0.3)+geom_ribbon(data = model_5_gg,aes(x=log(x),ymin=log(conf.low),ymax=log(conf.high),fill=group),alpha=0.1)+
  ylab("log(Shannon effective species)")+xlim(1,4)+
  xlab(bquote('log(Knoll surface area '~(m^2)))+
    theme(axis.text.x = element_text(size=20),
        axis.text.y = element_text(size =20),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        strip.text.x = element_text(size = 20))+
        scale_color_manual(values=fish_all,
                           name = "Year",
                           labels = c("Fall 2015", "Fall 2016",
                                      "Fall 2018","Spring 2020", "Fall 2020" ,
                                      "Spring 2021","Fall 2021"))+
        scale_fill_manual(values=fish_all,
                           name = "Year",
                           labels = c("Fall 2015", "Fall 2016",
                                      "Fall 2018","Spring 2020", "Fall 2020" ,
                                      "Spring 2021","Fall 2021"))


#ggsave("residence_fish_shannon_area.png" , width = 14, height = 5, dpi = 300,path = "C:\\Users\\TalG\\My Drive\\PHD\\chapter 1 strom effect\\R Scripts")

```


```{r}
model_2015<-lm(formula = log(shannon_effective) ~ log(Surface_Area)*Site +year_month +Max_depth ,data = wide_2015)
summary(model_2015)

sjPlot::plot_model(model_2015,type="eff",terms = c('year_month','Surface_Area'))+theme_classic()


surface_area = rep(seq(from = 1,to = 60,by = 1),24) # Length = 60
site = rep(unique(wide_2015$Site),each = 360) # Length = 4
year = rep(rep(unique(wide_2015$year_month),each = 60),4) # Length = 6

predict_no_out<-data.frame(Site = site,
                       year_month = year,
                       Surface_Area= surface_area,
                       Max_depth = rep(4.6))


predict_no_out$Surface_Area_log<-log(predict_no_out$Surface_Area)


predict_species_no_out<-predict(model_2015,predict_no_out,type="response")

predict_no_out<-cbind(predict_no_out,predict_species_no_out)

predict_no_out$site_year<-paste(predict_no_out$Site,predict_no_out$year_month,sep="_")

predict_no_out$species_number<-exp(predict_no_out$predict_species_no_out)


predict_no_out$year_month<- as.factor(predict_no_out$year_month)

l_type<-c("solid","solid","solid","twodash","solid","solid")

ggplot(data = predict_no_out,aes(x= Surface_Area,y = species_number,color= site))+
  theme_classic()+geom_smooth(alpha = 0.2)+
  geom_point(data = wide_2015,aes(Surface_Area,shannon_effective,color = Site),alpha = 0.3)+
  ylab("Shannon effective species")+
  xlab(bquote('Knoll surface area '~(m^2)))+
  theme(axis.text.x = element_text(size=20),
        axis.text.y = element_text(size =20),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        strip.text.x = element_text(size = 02))+
  scale_color_manual(name = "Site",values = color)




ggplot(data = predict_no_out,aes(x= Surface_Area,y = species_number,color= year_month))+
  theme_classic()+geom_smooth(alpha = 0.2)+   geom_point(data = wide_2015,aes(Surface_Area,shannon_effective,color = year_month),alpha = 0.3)+
  ylab("Shannon effective species")+
  xlab(bquote('Knoll surface area '~(m^2)))+
  theme(axis.text.x = element_text(size=20),
        axis.text.y = element_text(size =20),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        strip.text.x = element_text(size = 20))+
        scale_color_manual(values=fish_all,
                           name = "Year",
                           labels = c("Fall 2015", "Fall 2016",
                                      "Fall 2018","Spring 2020", "Fall 2020" ,
                                      "Spring 2021","Fall 2021"))



```

# models 2018-2021

```{r}
model_wide<-lm(formula = log(shannon_effective) ~ log(Surface_Area)*Site +year_month +Max_depth ,data = wide_data)
summary(model_2015)

sjPlot::plot_model(model_2015,type="eff",terms = c('year_month','Surface_Area'))+theme_classic()


surface_area = rep(seq(from = 1,to = 60,by = 1),24) # Length = 60
site = rep(unique(wide_2015$Site),each = 360) # Length = 4
year = rep(rep(unique(wide_2015$year_month),each = 60),4) # Length = 6

predict_no_out<-data.frame(Site = site,
                       year_month = year,
                       Surface_Area= surface_area,
                       Max_depth = rep(4.6))


predict_no_out$Surface_Area_log<-log(predict_no_out$Surface_Area)


predict_species_no_out<-predict(model_2015,predict_no_out,type="response")

predict_no_out<-cbind(predict_no_out,predict_species_no_out)

predict_no_out$site_year<-paste(predict_no_out$Site,predict_no_out$year_month,sep="_")

predict_no_out$species_number<-exp(predict_no_out$predict_species_no_out)


predict_no_out$year_month<- as.factor(predict_no_out$year_month)

l_type<-c("solid","solid","solid","twodash","solid","solid")

ggplot(data = predict_no_out,aes(x= Surface_Area,y = species_number,color= site))+
  theme_classic()+geom_smooth(alpha = 0.2)+
  geom_point(data = wide_data,aes(Surface_Area,shannon_effective,color = Site),alpha = 0.3)+
  ylab("Shannon effective species")+
  xlab(bquote('Knoll surface area '~(m^2)))+
  theme(axis.text.x = element_text(size=20),
        axis.text.y = element_text(size =20),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        strip.text.x = element_text(size = 02))+
  scale_color_manual(name = "Site",values = color)




ggplot(data = predict_no_out,aes(x= Surface_Area,y = species_number,color= year_month))+
  theme_classic()+geom_smooth(alpha = 0.2)+   geom_point(data = wide_data,aes(Surface_Area,shannon_effective,color = year_month),alpha = 0.3)+
  ylab("Shannon effective species")+
  xlab(bquote('Knoll surface area '~(m^2)))+
  theme(axis.text.x = element_text(size=20),
        axis.text.y = element_text(size =20),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        strip.text.x = element_text(size = 20))+
        scale_color_manual(values=fish_all,
                           name = "Year",
                           labels = c("Fall 2015", "Fall 2016",
                                      "Fall 2018","Spring 2020", "Fall 2020" ,
                                      "Spring 2021","Fall 2021"))
```
