==> 5.3 mobr 3dim.rmd <==
---
title: "5.3 mobr 3dim"
author: "TG"
date: "21 2 2022"
output:
   prettydoc::html_pretty:
    theme: cayman
    highlight: github
---

# Data wrangling


```{r}
load("fish_full_data.rdata")
load("survey_list.Rdata")
```

```{r message=FALSE, warning=FALSE}
library(dplyr)
library(tidyr)
library(lubridate)
library(ggplot2)
library(stringr)
library(Hmisc)
library(vegan)
library(mobr)
#library(sars)
library(lme4)
library(lmerTest)
library(MuMIn)
```

1) set one observer

```{r}
one_observer<-list()
loop<-1
set.seed(1)
for (i in unique(fish_full_data$survey_id)) {
  
  data <- fish_full_data %>% filter(survey_id == i)
  
  obs_amount <- length(unique(data$Observer))
  obs_name <- unique(data$Observer)
  chosen_obs <- sample(obs_name,1)  
 
  filter_data<-data %>% filter(Observer == chosen_obs)
  one_observer[[loop]]<-filter_data
  loop<-loop+1
  
  
}

one_observer<- bind_rows(one_observer)

rm(data,filter_data)
```


2) filter confidence = 0

for the years 2015-2017 where there is no confidance value i set the value 0

```{r}

one_observer$Confidence[is.na(one_observer$Confidence)]<-(0)
one_observer<-one_observer %>% filter(Confidence <1)

```

3) filter the north beach
```{r}

one_observer<-one_observer %>%
  filter(Site !="North beach",
         ! Knoll %in% c("Tamar","South bridge reserve","south bridge reserve",
                        "North bridge reserve","north bridge reserve"))
  
```



4) add the month to the sample

```{r include=FALSE}

one_observer$year_month<-ifelse(month(one_observer$Date)>6,
                                paste(one_observer$Year,"b",sep= " "),
                                paste(one_observer$Year,"a",sep= " "))


```




Create wide data for:

1. species by 2015 categories - wide_2015 (for each site comparison)

2. species by each year resolution (i.e Species column) - wide_data (for each year comparison)




## 1. mobr_2015_data

species by 2015 categories - wide_2015 (for each site comparison)


```{r include=FALSE}
all_fish_2015 <- one_observer %>%
  group_by(Site,Year,year_month,Knoll,DD_X,DD_Y,Surface_Area,Max_depth,Date,survey_id,Species_2015) %>%
  summarise("aboundance"=sum(Abundance))


wide_2015<-spread(all_fish_2015,Species_2015,aboundance,fill = 0)


```

## 2. mobr_data

species by each year resolution (i.e Species column) - wide_data (for each year comparison)


```{r include=FALSE}
all_fish <- one_observer %>%
  group_by(Site,Year,year_month,Knoll,DD_X,DD_Y,Surface_Area,Max_depth,Date,survey_id,Species) %>%
  summarise("aboundance"=sum(Abundance))

after_storm_knoll<- one_observer %>% filter(year_month =="2020 a") %>%drop_na(DD_X) %>% distinct(knoll_id) 


wide_data <- spread(all_fish,Species,aboundance,fill = 0)

wide_data <- wide_data %>%
  mutate("Site_Knoll" = paste(Site,Knoll,sep="_"),.after = Knoll)





```

I used a subset of all knoll to represent only knolls that were surveyd in spring 2020 (after the storm)

"C:\\Users\\TalG\\My Drive\\PHD\\chapter 1 strom effect\\R Scripts\\plots for ms\\mobr"

# compere 2018 to storm (spring 2020)

## with schools

I got 51 knoll that have coords and were sampled in both periods

```{r}
data.1 <- wide_data %>%
  filter(year_month %in% c("2018 b","2020 a")) %>%  filter(Site_Knoll %in% after_storm_knoll$knoll_id) 
  
species.data.1<-data.1[,12:ncol(data.1)]
site.data.1<-data.1[,1:11]

mobr_data.1 <- make_mob_in(species.data.1, site.data.1,coord_names = c("DD_Y","DD_X"))

min(rowSums(species.data.1))
min(colSums(species.data.1))

# rarefaction curves

plot_rarefaction(mobr_data.1, 'year_month',method=  'IBR', lwd=4)
plot_rarefaction(mobr_data.1, 'year_month',method=  'SBR', lwd=4)
plot_rarefaction(mobr_data.1, 'year_month',method=  'sSBR', lwd=4)
plot_abu(mobr_data.1, 'year_month', type='rad', pooled=T, log='x') 

# 2 dim stats

site_stats.1 <- get_mob_stats(mobr_data.1, group_var = "year_month",n_perm = 100,boot_groups = TRUE,effort_samples = 50)

plot(site_stats.1)


# 3 dim

delta.1 <- get_delta_stats(mobr_data.1,env_var = "year_month",type="discrete",n_perm = 20)

plot(delta.1,stat = "b1")

```

```{r}
pdf(file="plots for ms\\mobr\\with_school_2018_2020a.pdf",width=10, height=7)
plot(delta.1)
dev.off()
```

## without schools


```{r}
data.2 <- wide_data %>%
  filter(year_month %in% c("2018 b","2020 a"))  %>%
  drop_na(DD_X) %>%
  filter(Site_Knoll %in% after_storm_knoll$knoll_id)

data.2$`Chromis viridis`<-NULL
data.2$`Pseudanthias squamipinnis`<-NULL
data.2$`Neopomacentrus miryae`<-NULL
data.2$`Spratelloides gracilis`<-NULL
  
species.data.2<-data.2[,12:ncol(data.2)]
site.data.2<-data.2[,1:11]

mobr_data.2 <- make_mob_in(species.data.2, site.data.2,coord_names = c("DD_Y","DD_X"))

# rarefaction curves

plot_rarefaction(mobr_data.2, 'year_month',method=  'IBR', lwd=4)
plot_rarefaction(mobr_data.2, 'year_month',method=  'SBR', lwd=4)
plot_rarefaction(mobr_data.2, 'year_month',method=  'sSBR', lwd=4)
plot_abu(mobr_data.2, 'year_month', type='rad', pooled=T, log='x') 

# 2 dim stats

site_stats.2 <- get_mob_stats(mobr_data.2, group_var = "year_month",n_perm = 100,boot_groups = TRUE,effort_samples = 50)

plot(site_stats.2)


# 3 dim

delta.2 <- get_delta_stats(mobr_data.2,env_var = "year_month",type="discrete",n_perm = 20)

plot(delta.2,stat = "b1")


```

```{r}
pdf(file="plots for ms\\mobr\\without_school_2018_2020a.pdf",width=10, height=7)
plot(delta.2)
dev.off()
```

# compare 2018 to fall 2020 

## with schools


```{r}
data.3 <- wide_data %>%
  filter(year_month %in% c("2018 b","2020 b"))  %>%
  drop_na(DD_X) %>%
  filter(Site_Knoll %in% after_storm_knoll$knoll_id) %>% 
  filter(Site_Knoll != "Princess_6")


data.3_test<-data.3 %>% group_by(Site_Knoll) %>% summarise(n=n())
  
species.data.3<-data.3[,12:ncol(data.3)]
site.data.3<-data.3[,1:11]

mobr_data.3 <- make_mob_in(species.data.3, site.data.3,coord_names = c("DD_Y","DD_X"))

# rarefaction curves

plot_rarefaction(mobr_data.3, 'year_month',method=  'IBR', lwd=4)
plot_rarefaction(mobr_data.3, 'year_month',method=  'SBR', lwd=4)
plot_rarefaction(mobr_data.3, 'year_month',method=  'sSBR', lwd=4)
plot_abu(mobr_data.3, 'year_month', type='rad', pooled=T, log='x') 

# 2 dim stats

site_stats.3 <- get_mob_stats(mobr_data.3, group_var = "year_month",n_perm = 100,boot_groups = TRUE,effort_samples = 50)

plot(site_stats.3)


# 3 dim

delta.3 <- get_delta_stats(mobr_data.3,env_var = "year_month",type="discrete",n_perm = 10)

plot(delta.3,stat = "b1")

```

```{r}
pdf(file="plots for ms\\mobr\\with_school_2018_2020a.pdf",width=10, height=7)
plot(delta.3)
dev.off()
```



## without schools




```{r}
data.4 <- wide_data %>%
  filter(year_month %in% c("2018 b","2020 b"))  %>%
  filter(Site_Knoll %in% after_storm_knoll$knoll_id)%>%
  drop_na(DD_X) %>% 
  filter(Site_Knoll != "Princess_6")
  

data.4$`Chromis viridis`<-NULL
data.4$`Pseudanthias squamipinnis`<-NULL
data.4$`Neopomacentrus miryae`<-NULL
data.4$`Spratelloides gracilis`<-NULL
  
species.data.4<-data.4[,12:ncol(data.4)]
site.data.4<-data.4[,1:11]

mobr_data.4 <- make_mob_in(species.data.4,
                           site.data.4,
                           coord_names = c("DD_Y","DD_X"))

# rarefaction curves

plot_rarefaction(mobr_data.4, 'year_month',method=  'IBR', lwd=4)
plot_rarefaction(mobr_data.4, 'year_month',method=  'SBR', lwd=4)
plot_rarefaction(mobr_data.4, 'year_month',method=  'sSBR', lwd=4)
plot_abu(mobr_data.4, 'year_month', type='rad', pooled=T, log='x') 

# 2 dim stats

site_stats.4 <- get_mob_stats(mobr_data.4, group_var = "year_month",n_perm = 100,boot_groups = TRUE,effort_samples = 50)

plot(site_stats.4)


# 3 dim

delta.4 <- get_delta_stats(mobr_data.4,env_var = "year_month",type="discrete",n_perm = 20,extrapolate = T)

plot(delta.4,stat = "b1")

min(rowSums(species.data.4))
```

```{r}
pdf(file="plots for ms\\mobr\\without_school_2018_2020b.pdf",width=10, height=7)
plot(delta.4)
dev.off()
```

# spring 2020 (storm) to fall 2021

## with schools
## without schools

# appendix 2 - fall 2018 to spring 2021

## with schools
## without schools
