==> 15.1 compare fish density.rmd <==
---
title: "15.1 compare fish density"
author: "TG"
date: "23 12 2021"
output:
   prettydoc::html_pretty:
    theme: cayman
    highlight: github
---


```{r message=FALSE, warning=FALSE}
library(dplyr)
library(tidyr)
library(lubridate)
library(ggplot2)
library(stringr)
library(Hmisc)
library(vegan)
library(mobr)
library(lme4)
library(lmerTest)
library(MuMIn)
library(Rarefy)
library(rareNMtests)
library(plotrix)
```



```{r}
load("fish_full_data.rdata")
color<-c("#56dde0", "#33658a", "#f5af2d","#f2641a")
fish_all<-c("#669bbc","#336683","#003049","#ed3140","#c1121f","#710912")
fish_k<-c("#669bbc","#336683","#1a4b66","#003049","#ed3140","#c1121f","#710912")

```

1) set one observer

```{r}
one_observer<-list()
loop<-1
set.seed(1)
for (i in unique(fish_full_data$survey_id)) {
  
  data <- fish_full_data %>% filter(survey_id == i)
  
  obs_amount <- length(unique(data$Observer))
  obs_name <- unique(data$Observer)
  chosen_obs <- sample(obs_name,1)  
 
  filter_data<-data %>% filter(Observer == chosen_obs)
  one_observer[[loop]]<-filter_data
  loop<-loop+1
  
  
}

one_observer<- bind_rows(one_observer)

rm(data,filter_data)
```




3) filter the north beach
```{r}
one_observer<-one_observer %>%
  filter(Site !="North beach",
         ! Knoll %in% c("Tamar","South bridge reserve","south bridge reserve",
                        "North bridge reserve","north bridge reserve"))
                                      
```



4) add the month to the sample

```{r include=FALSE}

one_observer$year_month<-ifelse(month(one_observer$Date)>6,
                                paste(one_observer$Year,"b",sep= " "),
                                paste(one_observer$Year,"a",sep= " "))


```

# compare the number of schools between the sites


```{r}

schools<- one_observer %>% filter(Species_2015 %in% c("Pseudanthias spp.","Chromis viridis","Neopomacentrus miryae")) %>% group_by(Site,year_month) %>% summarise("school"=sum(Abundance))

ggplot(data = schools,aes(x=year_month,y=school,color=Site))+geom_point()


ggplot(data = schools,aes(x=Site,y=school,fill=Site))+geom_boxplot()+
  xlab("Site")+ ggtitle("Schooling fish abundance")+
  ylab("Number of individuals")+ theme_bw()+
    theme(axis.text.x = element_text(size=20),
        plot.title = element_text(size=24),
        axis.text.y = element_text(size =20),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        legend.position = "none",
        strip.text.x = element_text(size = 20))+
        scale_fill_manual(values= color)

#ggsave("school_fish_abundances.png" , width = 10, height = 8, dpi = 300,path = "C:\\Users\\TalG\\Documents\\PhD\\Eco camp\\R Scripts\\plots nov 2021")

```
# Schools per m2

```{r}
schools_per_m<- one_observer %>% filter(Species_2015 %in% c("Pseudanthias spp.","Chromis viridis","Neopomacentrus miryae")) %>% group_by(Site,year_month,Knoll,Surface_Area) %>% summarise("inds"=sum(Abundance)) %>% mutate("inds_m"=inds/Surface_Area)


schools_per_m<-schools_per_m %>% group_by(Site) %>% summarise("mean_density"=mean(inds_m),"se"=std.error(inds_m),n=n())

ggplot(data = schools_per_m,aes(x=Site,y=mean_density,fill = Site))+
geom_bar(stat="identity")+geom_errorbar(aes(ymin=mean_density-se,ymax=mean_density+se),width =0.1)+
  scale_fill_manual(values = color)+
  geom_text(aes(label = n,y=0.9), size = 7)+  
  xlab("Site")+ ggtitle("Fish abundance ( schools)")+
  ylab(bquote('Mean schools density '(m^2)))+
  theme_bw()+
    theme(axis.text.x = element_text(size=20),
        plot.title = element_text(size=24),
        axis.text.y = element_text(size =20),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        legend.position = "none",
        strip.text.x = element_text(size = 20))
  
ggsave("school_density_m2.png" , width = 10, height = 8, dpi = 300,path = "C:\\Users\\TalG\\Documents\\PhD\\Eco camp\\R Scripts\\plots nov 2021")
```

# Schools per m2 by year


```{r}
schools_per_m<- one_observer %>% filter(Species_2015 %in% c("Pseudanthias spp.","Chromis viridis","Neopomacentrus miryae")) %>% group_by(Site,year_month,Knoll,Surface_Area) %>% summarise("inds"=sum(Abundance)) %>% mutate("inds_m"=inds/Surface_Area)


schools_per_m<-schools_per_m %>% group_by(year_month) %>% summarise("mean_density"=mean(inds_m),"se"=std.error(inds_m),n=n())

ggplot(data = schools_per_m,aes(x=year_month,y=mean_density,fill = year_month))+
geom_bar(stat="identity")+
  geom_errorbar(aes(ymin=mean_density-se,ymax=mean_density+se),width =0.1)+
  scale_fill_manual(values = fish_k)+
  geom_text(aes(label = n,y=0.9), size = 7,color="White")+  
  xlab("Sample")+ ggtitle("Fish abundance ( schools)")+
  ylab(bquote('Mean schools density '(m^2)))+
  theme_bw()+
    theme(axis.text.x = element_text(size=20),
        plot.title = element_text(size=24),
        axis.text.y = element_text(size =20),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        legend.position = "none",
        strip.text.x = element_text(size = 20))+
    scale_x_discrete(labels = c("Fall 2015", "Fall 2016","Fall 2017",
                                      "Fall 2018","Spring 2020",
                                      "Fall 2020" ,
                                      "Spring 2021"))
  
ggsave("school_density_m2_year.png" , width = 10, height = 8, dpi = 300)




schools_per_m<- one_observer %>% filter(Species_2015 %in% c("Pseudanthias spp.","Chromis viridis","Neopomacentrus miryae")) %>% group_by(Site,year_month,Knoll,Surface_Area) %>% summarise("inds"=sum(Abundance)) %>% mutate("inds_m"=inds/Surface_Area)


schools_per_m<-schools_per_m %>% group_by(year_month,Site) %>% summarise("mean_density"=mean(inds_m),"se"=std.error(inds_m),n=n())

ggplot(data = schools_per_m,aes(x=year_month,y=mean_density,fill = year_month))+
geom_bar(stat="identity")+facet_wrap(.~ Site)+
  geom_errorbar(aes(ymin=mean_density-se,ymax=mean_density+se),width =0.1)+
  scale_fill_manual(values = fish_k)+
  geom_text(aes(label = n,y=35), size = 7)+  
  xlab("Sample")+ ggtitle("Fish abundance ( schools)")+
  ylab(bquote('Mean schools density '(m^2)))+
  theme_bw()+
    theme(axis.text.x = element_text(size=20),
        plot.title = element_text(size=24),
        axis.text.y = element_text(size =20),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        legend.position = "none",
        strip.text.x = element_text(size = 20))+
    scale_x_discrete(labels = c("Fall 2015", "Fall 2016","Fall 2017",
                                      "Fall 2018","Spring 2020",
                                      "Fall 2020" ,
                                      "Spring 2021"))
  
ggsave("school_density_m2_year_site.png" , width = 10, height = 8, dpi = 300)





```




# compare the total number of individuals between the sites


```{r}

ind<- one_observer %>% group_by(Site,year_month) %>% summarise("abundance"=sum(Abundance))


ggplot(data = ind,aes(x=Site,y=abundance,fill=Site))+geom_boxplot()+
  xlab("Site")+ ggtitle("Fish abundance")+
  ylab("Number of individuals")+ theme_bw()+
    theme(axis.text.x = element_text(size=20),
        plot.title = element_text(size=24),
        axis.text.y = element_text(size =20),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        legend.position = "none",
        strip.text.x = element_text(size = 20))+
        scale_fill_manual(values= color)

ggsave("fish_abundances.png" , width = 10, height = 8, dpi = 300,path = "C:\\Users\\TalG\\Documents\\PhD\\Eco camp\\R Scripts\\plots nov 2021")

```

# ind no schools

```{r}
ind_no_school<- one_observer %>% filter(!Species_2015 %in% c("Pseudanthias spp.","Chromis viridis","Neopomacentrus miryae")) %>% group_by(Site,year_month) %>% summarise("inds"=sum(Abundance))



ggplot(data = ind_no_school,aes(x=Site,y=inds,fill=Site))+geom_boxplot()+
  xlab("Site")+ ggtitle("Fish abundance (no schools)")+
  ylab("Number of individuals")+ theme_bw()+
    theme(axis.text.x = element_text(size=20),
        plot.title = element_text(size=24),
        axis.text.y = element_text(size =20),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        legend.position = "none",
        strip.text.x = element_text(size = 20))+
        scale_fill_manual(values= color)

ggsave("fish_abundances_no_school.png" , width = 10, height = 8, dpi = 300,path = "C:\\Users\\TalG\\Documents\\PhD\\Eco camp\\R Scripts\\plots nov 2021")

```

# fish abundance(no schools) per m2

```{r}
ind_per_m<- one_observer %>% filter(!Species_2015 %in% c("Pseudanthias spp.","Chromis viridis","Neopomacentrus miryae")) %>% group_by(Site,year_month,Knoll,Surface_Area) %>% summarise("inds"=sum(Abundance)) %>% mutate("inds_m"=inds/Surface_Area)


ind_per_m<-ind_per_m %>% group_by(Site) %>% summarise("mean_density"=mean(inds_m),"se"=std.error(inds_m),n=n())

ggplot(data = ind_per_m,aes(x=Site,y=mean_density,fill = Site))+
geom_bar(stat="identity")+geom_errorbar(aes(ymin=mean_density-se,ymax=mean_density+se),width =0.1)+
  scale_fill_manual(values = color)+
  geom_text(aes(label = n,y=0.5), size = 7)+  
  xlab("Site")+ ggtitle("Fish abundance (no schools)")+
  ylab(bquote('Mean fish density '(m^2)))+
  theme_bw()+
    theme(axis.text.x = element_text(size=20),
        plot.title = element_text(size=24),
        axis.text.y = element_text(size =20),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        legend.position = "none",
        strip.text.x = element_text(size = 20))
  
ggsave("fish_density_m2.png" , width = 10, height = 8, dpi = 300,path = "C:\\Users\\TalG\\Documents\\PhD\\Eco camp\\R Scripts\\plots nov 2021")

```


# fish per m2 - for each year

```{r}
ind_per_m_year<- one_observer %>% filter(!Species_2015 %in% c("Pseudanthias spp.","Chromis viridis","Neopomacentrus miryae")) %>% filter(year_month!="2017 b")%>%  group_by(Site,year_month,Knoll,Surface_Area) %>% summarise("inds"=sum(Abundance)) %>% mutate("inds_m"=inds/Surface_Area)


ind_per_m_year<-ind_per_m_year %>% group_by(year_month) %>% summarise("mean_density"=mean(inds_m),"se"=std.error(inds_m),n=n())

ind_per_m_year$year_month<-as.factor(ind_per_m_year$year_month)

ggplot(data = ind_per_m_year,aes(x=year_month,y=mean_density,fill = year_month))+
geom_bar(stat="identity")+geom_errorbar(aes(ymin=mean_density-se,ymax=mean_density+se),width =0.1)+
  geom_text(aes(label = n,y=0.5), size = 7,color= "white")+  
  xlab("Year")+ ggtitle("Fish abundance (no schools)")+
  ylab(bquote('Mean fish density '(m^2)))+
  theme_bw()+
    theme(axis.text.x = element_text(size=20),
        plot.title = element_text(size=24),
        axis.text.y = element_text(size =20),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        legend.position = "none",
        strip.text.x = element_text(size = 20))+
  scale_fill_manual(values= fish_all,
                           name = "Year",
                           labels = c("Fall 2015", "Fall 2016",
                                      "Fall 2018","Spring 2020",
                                      "Fall 2020" ,
                                      "Spring 2021"))+
  scale_x_discrete(labels = c("Fall 2015", "Fall 2016",
                                      "Fall 2018","Spring 2020",
                                      "Fall 2020" ,
                                      "Spring 2021"))
  
ggsave("fish_density_m2_year.png" , width = 10, height = 8, dpi = 300,path = "C:\\Users\\TalG\\Documents\\PhD\\Eco camp\\R Scripts\\plots nov 2021")

```




# compare the number of ind per knoll

to avoid bias of  different number of knolls or area in the sites
ill do it without schools


```{r}
ind_no_school_knoll<- one_observer %>% filter(!Species_2015 %in% c("Pseudanthias spp.","Chromis viridis","Neopomacentrus miryae")) %>% group_by(Site,year_month,knoll_id) %>% summarise("inds"=sum(Abundance))


ggplot(data = ind_no_school_knoll,aes(x=Site,y=inds,fill=Site))+
  geom_boxplot(notch = F)+
  xlab("Site")+ ggtitle("Fish abundance (no schools)")+
  ylab("Individuals/knoll")+ theme_bw()+
    theme(axis.text.x = element_text(size=20),
        plot.title = element_text(size=24),
        axis.text.y = element_text(size =20),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        legend.position = "none",
        strip.text.x = element_text(size = 20))+
        scale_fill_manual(values= color)+ylim(0,500)

ggsave("fish_abundances_no_school_KNOLL.png" , width = 10, height = 8, dpi = 300,path = "C:\\Users\\TalG\\Documents\\PhD\\Eco camp\\R Scripts\\plots nov 2021")


```



# compare the number of fishes along the years

for this ill use for each site only the knolls that were samples after the storm

```{r}
after_storm_knoll<- one_observer %>% filter(year_month =="2020 a") %>% distinct(knoll_id)


after_storm_data<-one_observer %>% filter(knoll_id %in% after_storm_knoll$knoll_id) %>%
  group_by(Site,year_month) %>%
  summarise("inds" = sum(Abundance))


ggplot(data = after_storm_data,aes(x=Site,y=inds,color=year_month))+geom_point(size = 3.5)+ggtitle("Fish abundance")+
  ylab("Number of individuals")+ theme_bw()+
    theme(axis.text.x = element_text(size=20),
        plot.title = element_text(size=24),
        axis.text.y = element_text(size =20),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        strip.text.x = element_text(size = 20))+
        scale_color_manual(values= fish_k,
                           name = "Year",
                           labels = c("Fall 2015", "Fall 2016","Fall 2017",
                                      "Fall 2018","Spring 2020", "Fall 2020" ,
                                      "Spring 2021","Fall 2021"))

ggsave("only_storm_knoll_ind_fish.png" , width = 10, height = 8, dpi = 300,path = "C:\\Users\\TalG\\Documents\\PhD\\Eco camp\\R Scripts\\plots nov 2021")

```
 
 
 
# compare the number of fishes along the years withput schools

for this ill use for each site only the knolls that were samples after the storm

```{r}



after_storm_data_ns<-one_observer %>%
  filter(knoll_id %in% after_storm_knoll$knoll_id) %>%
  filter(!Species_2015 %in% c("Pseudanthias spp.","Chromis viridis","Neopomacentrus miryae")) %>% 
  group_by(Site,year_month) %>%
  summarise("inds" = sum(Abundance))


ggplot(data = after_storm_data_ns,aes(x=Site,y=inds,color=year_month))+geom_point(size = 3.5)+ggtitle("Fish abundance (no schools)")+
  ylab("Number of individuals")+ theme_bw()+
    theme(axis.text.x = element_text(size=20),
        plot.title = element_text(size=24),
        axis.text.y = element_text(size =20),
        axis.title.x = element_text(size=22),
        axis.title.y = element_text(size=22),
        legend.title = element_text(size=22),
        legend.text = element_text(size=20),
        strip.text.x = element_text(size = 20))+
        scale_color_manual(values= fish_k,
                           name = "Year",
                           labels = c("Fall 2015", "Fall 2016","Fall 2017",
                                      "Fall 2018","Spring 2020", "Fall 2020" ,
                                      "Spring 2021","Fall 2021"))

ggsave("only_storm_knoll_ind_fish_no_schools.png" , width = 10, height = 8, dpi = 300,path = "C:\\Users\\TalG\\Documents\\PhD\\Eco camp\\R Scripts\\plots nov 2021")

```
